backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "ansible-core>=2.20.3",
    "arq>=0.25.0",
    "asyncssh>=2.22.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "redis>=7.2.1",
    "ruamel-yaml>=0.19.1",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
